pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "https://github.com/Parker-idc/python-project-50#description-"
authors = ["Parker-idc <egor.kruchenkov@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Parker-idc/python-project-50"
classifiers = [
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.10",
    "Topic :: Terminals :: Games"
]
packages = [{include = "gendiff"}]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
pytest = "^7.2.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
